Bump to version 1.0, introduce nullable array handling
Resolves #4

Add tests to ensure the `nullIfEmpty` function behaves correctly,
returning null when expected.

I’ve also added integration tests using an sqlite database to ensure
that the functionality is correctly called in combination with
Laravel’s Eloquent.

It does not appear as though this was a simple matter of checking if
the passed in `$value` is an array and determining if it was `empty`,
given that any value that is cast to an `array` will be done *before*
being passed through the `saving` observer.

This means checking if the key `isJsonCastable` and checking its value
is empty.